api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
cachetools==4.1.0         # via google-auth
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
google-api-core[grpc]==1.16.0  # via google-cloud-core, google-cloud-firestore
google-auth==1.13.1       # via google-api-core
google-cloud-core==1.3.0  # via google-cloud-firestore
google-cloud-firestore==1.6.2  # via -r requirements.in
googleapis-common-protos==1.51.0  # via google-api-core
grpcio==1.28.1            # via google-api-core
idna==2.9                 # via requests
meetup-api==0.1.1         # via -r requirements.in
protobuf==3.11.3          # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pytz==2019.3              # via google-api-core, google-cloud-firestore
requests==2.23.0          # via google-api-core, meetup-api
rsa==4.0                  # via google-auth
six==1.14.0               # via google-api-core, google-auth, grpcio, meetup-api, protobuf
urllib3==1.25.8           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
